Add participants, delete comments

+++ base/views.py
@@ -9,7 +9,7 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.forms import UserCreationForm
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
@@ -87,7 +87,17 @@ def home(request):
 
 def room(request, pk):
     room =Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', room.id)
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     return render(request, "base/room.html", context)
 
 
@@ -122,7 +132,22 @@ def update_room(request, pk):
 @login_required(login_url="/login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != r.host:
+        return HttpResponse("You are not allowed here - not your room.")
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
+
+@login_required(login_url="/login")
+def delete_message(request, pk, rPk):
+   # room = Room.objects.get(id=rPk)
+    message = Message.objects.get(id=pk)
+    if request.user !=  message.user:
+        return HttpResponse("You are not allowed here ")
+
+    if request.method=='POST':
+        message.delete()
+        return redirect('room', rPk)
+    return render(request, 'base/delete.html', {'obj': message})
